events:btn
btn.onclick: When the button is clicked, its background color changes to red, and its width and height adjust.

btn.ondblclick: When the button is double-clicked, the background turns green, and its dimensions are changed.

window.onmouseover: When the mouse hovers over the window, the background color of the (not the button) changes to green.window

window.onload: When the page is fully loaded, the button's background color is set to orange.